% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacenOracle.R
\name{roracle}
\alias{roracle}
\title{Function that copies a dataframe from R to Oracle, creating SQL structure}
\usage{
roracle(dataframe, tableName,
  schemaName = getOption("bacen.oracle.schemaName"),
  hostName = getOption("bacen.oracle.hostName"),
  port = getOption("bacen.oracle.port"),
  sid = getOption("bacen.oracle.sid"),
  user = getOption("bacen.oracle.user"),
  password = getOption("bacen.oracle.password"), useSQLldr = TRUE)
}
\arguments{
\item{tableName}{tableName to be created in database}

\item{schemaName}{schemaName - settable via options("bacen.oracle.schemaName") - default: manut_shinny}

\item{hostName}{hostname - settable via options("bacen.oracle.hostName")}

\item{port}{port number - settable via options("bacen.oracle.port")}

\item{sid}{sid - settable via options("bacen.oracle.sid")}

\item{user}{username - settable via options("bacen.oracle.user")}

\item{password}{password - settable via options("bacen.oracle.password")}

\item{useSQLldr}{useSQLldr - indicate if the user prefer to use SQLldr method}

\item{df}{dataframe object to be transfered}
}
\description{
Function that copies a dataframe from R to Oracle, creating SQL structure
}
\examples{
\dontrun{
df <- data.frame(coluna1 = c(1,2,3), coluna2 = c("a","b","c"))
roracle(df, "tabela", "schema")
}
}
