% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacenVertica.R
\name{verticaGetQuery}
\alias{verticaGetQuery}
\title{Function that executes a SQL query that result in a dataframe (select registers from a table)}
\usage{
verticaGetQuery(query, hostName = getOption("bacen.vertica.hostName"),
  port = getOption("bacen.vertica.port"),
  databaseName = getOption("bacen.vertica.databaseName"),
  user = getOption("bacen.vertica.user"),
  password = getOption("bacen.vertica.password"))
}
\arguments{
\item{query}{query to be executed}

\item{hostName}{hostname - settable via options("bacen.vertica.hostName")}

\item{port}{port number - settable via options("bacen.vertica.port")}

\item{databaseName}{name - settable via options("bacen.vertica.databaseName")}

\item{user}{username - settable via options("bacen.vertica.user")}

\item{password}{password - settable via options("bacen.vertica.password")}
}
\value{
dataframe with resultset
}
\description{
Function that executes a SQL query that result in a dataframe (select registers from a table)
}
\examples{
\dontrun{
verticaGetQuery("SELECT 1")
}
}
