% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacenVertica.R
\name{verticaSendUpdate}
\alias{verticaSendUpdate}
\title{Function that executes a SQL query that result in a modification (update, alter, drop, etc)}
\usage{
verticaSendUpdate(query, hostName = getOption("bacen.vertica.hostName"),
  port = getOption("bacen.vertica.port"),
  databaseName = getOption("bacen.vertica.databaseName"),
  user = getOption("bacen.vertica.user"),
  password = getOption("bacen.vertica.password"))
}
\arguments{
\item{query}{query to be executed}

\item{hostName}{hostname - settable via options("bacen.vertica.hostName")}

\item{port}{port number - settable via options("bacen.vertica.port")}

\item{databaseName}{name - settable via options("bacen.vertica.databaseName")}

\item{user}{username - settable via options("bacen.vertica.user")}

\item{password}{password - settable via options("bacen.vertica.password")}
}
\description{
Function that executes a SQL query that result in a modification (update, alter, drop, etc)
}
\examples{
\dontrun{
verticaGetQuery("UPDATE tabela SET campo = 'valor'")
}
}
