% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacen.R
\name{rertica}
\alias{rertica}
\title{Function that copy a dataframe from R to Vertica, creating SQL structure}
\usage{
rertica(host, user, password, schemaName, df, tableName)
}
\arguments{
\item{host}{hostname of database or ip from database}

\item{user}{username to login in database}

\item{password}{password to login in database}

\item{schemaName}{schemaname in database to transfer dataframe in}

\item{df}{dataframe object to be transfered}

\item{tableName}{tablename to be created in database}
}
\description{
Function that copy a dataframe from R to Vertica, creating SQL structure
}
